optimize-services&queries add author checks in related methods

+++ AI-Social-Platform.Services.Data/PublicationService.cs
@@ -2,32 +2,34 @@
 using AI_Social_Platform.Data.Models.Publication;
 using AI_Social_Platform.Services.Data.Interfaces;
 using AI_Social_Platform.Services.Data.Models.PublicationDtos;
+using Microsoft.AspNetCore.Http;
 using Microsoft.EntityFrameworkCore;
+using System.Security.Claims;
 
 namespace AI_Social_Platform.Services.Data;
 
 public class PublicationService : IPublicationService
 {
     private readonly ASPDbContext dbContext;
+    private readonly HttpContext httpContext;
 
-    public PublicationService(ASPDbContext dbContext)
+    public PublicationService(ASPDbContext dbContext, IHttpContextAccessor accessor)
     {
         this.dbContext = dbContext;
+        this.httpContext = accessor.HttpContext!;
     }
     public async Task<IEnumerable<PublicationDto>> GetPublicationsAsync()
     {
         var publications = await
             dbContext.Publications
-                .OrderByDescending(p => p.DateCreated)
                 .Select(p => new PublicationDto()
                 {
                     Id = p.Id,
                     Content = p.Content,
-                    DateCreated = p.DateCreated.ToShortDateString() 
-                                  + " " 
-                                  + p.DateCreated.ToShortTimeString(),
+                    DateCreated = p.DateCreated,
                     AuthorId = p.AuthorId
                 })
+                .OrderByDescending(p => p.DateCreated)
                 .ToListAsync();
 
         return publications;
@@ -48,9 +50,7 @@ public class PublicationService : IPublicationService
         {
             Id = publication.Id,
             Content = publication.Content,
-            DateCreated = publication.DateCreated.ToShortDateString()
-                          + " " 
-                          + publication.DateCreated.ToShortTimeString(),
+            DateCreated = publication.DateCreated,
             AuthorId = publication.AuthorId
         };
 
@@ -60,11 +60,12 @@ public class PublicationService : IPublicationService
     
     public async Task CreatePublicationAsync(PublicationFormDto dto)
     {
-       Publication publication = new()
-       {
+        var userId = await GetUserId();
+        Publication publication = new()
+        {
            Content = dto.Content,
-           AuthorId = dto.AuthorId
-       };
+           AuthorId = userId
+        };
 
        await dbContext.AddAsync(publication);
        await dbContext.SaveChangesAsync();
@@ -73,13 +74,14 @@ public class PublicationService : IPublicationService
     public async Task UpdatePublicationAsync(PublicationFormDto dto, Guid id)
     {
         var publication = await dbContext.Publications.FirstOrDefaultAsync(p => p.Id == id);
+        var userId = await GetUserId();
 
         if (publication == null)
         {
             throw new Exception("Publication not found");
         }
 
-        if (publication.AuthorId != dto.AuthorId)
+        if (publication.AuthorId != userId)
         {
             throw new Exception("You are not the author of this publication");
         }
@@ -91,15 +93,22 @@ public class PublicationService : IPublicationService
     public async Task DeletePublicationAsync(Guid id)
     {
         var publication = await dbContext.Publications
-            .Include(p => p.Comments)
+            //.Include(p => p.Comments)
             .FirstOrDefaultAsync(p => p.Id == id);
+
+        var userId = await GetUserId();
 
         if (publication == null)
         {
             throw new Exception("Publication not found");
         }
 
-        dbContext.Comments.RemoveRange(publication.Comments);
+        if (publication.AuthorId != userId)
+        {
+            throw new Exception("You are not the author of this publication");
+        }
+
+        //dbContext.Comments.RemoveRange(publication.Comments);
         dbContext.Publications.Remove(publication);
         await dbContext.SaveChangesAsync();
     }
@@ -109,6 +118,7 @@ public class PublicationService : IPublicationService
     {
         var publication = await dbContext.Publications
             .FirstOrDefaultAsync(p => p.Id == publicationId);
+        var userId = await GetUserId();
 
         if (publication == null)
         {
@@ -117,7 +127,7 @@ public class PublicationService : IPublicationService
 
         await dbContext.Comments.AddAsync(new Comment()
         {
-            AuthorId = dto.AuthorId,
+            AuthorId = userId,
             Content = dto.Content,
             PublicationId = publicationId
         });
@@ -127,13 +137,14 @@ public class PublicationService : IPublicationService
     public async Task UpdateCommentAsync(CommentFormDto dto, Guid id)
     {
        var comment = await dbContext.Comments.FirstOrDefaultAsync(c => c.Id == id);
+       var userId = await GetUserId();
 
-       if (comment == null)
+        if (comment == null)
        {
            throw new Exception("Comment not found");
        }
 
-       if (comment.AuthorId != dto.AuthorId)
+       if (comment.AuthorId != userId)
        {
            throw new Exception("You are not the author of this comment");
        }
@@ -145,10 +156,16 @@ public class PublicationService : IPublicationService
     public async Task DeleteCommentAsync(Guid id)
     {
         var comment = await dbContext.Comments.FirstOrDefaultAsync(c => c.Id == id);
+        var userId = await GetUserId();
 
-        if(comment == null)
+        if (comment == null)
         {
             throw new Exception("Comment not found");
+        }
+
+        if (comment.AuthorId != userId)
+        {
+            throw new Exception("You are not the author of this comment");
         }
 
         dbContext.Comments.Remove(comment);
@@ -159,19 +176,30 @@ public class PublicationService : IPublicationService
     {
         var publicationComments = await dbContext.Comments
             .Where(c => c.PublicationId == publicationId)
-            .OrderByDescending(c => c.DateCreated)
             .Select(c => new CommentDto()
             {
                 Id = c.Id,
                 Content = c.Content,
-                DateCreated = c.DateCreated.ToShortDateString() 
-                              + " " 
-                              + c.DateCreated.ToShortTimeString(),
+                DateCreated = c.DateCreated,
                 AuthorId = c.AuthorId,
                 PublicationId = c.PublicationId
             })
+            .OrderByDescending(c => c.DateCreated)
             .ToListAsync();
 
         return publicationComments;
+    }
+
+    private async Task<Guid> GetUserId()
+    {
+        var userEmail = httpContext.User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.NameIdentifier)?.Value!;
+        var userId = await dbContext.Users.FirstOrDefaultAsync(u => u.Email == userEmail);
+
+        if (userId == null)
+        {
+            throw new Exception("User not found");
+        }
+
+        return userId.Id;
     }
 }
